# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chayn Survivor AI API",
    "version": "1.0.0",
    "description": "API endpoints for the Chayn Survivor AI application, including policy validation, data collection, and letter generation."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/policy-validation/process-document": {
      "post": {
        "summary": "Process next document in validation session",
        "description": "Processes the next document in the validation queue. Uses AI to analyze the document against current policies and determine if updates are needed. Call repeatedly until all documents are processed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "validationId": {
                    "type": "string",
                    "description": "The validation session ID returned from the initialize endpoint."
                  }
                },
                "required": ["validationId"]
              },
              "example": {
                "validationId": "validation_1234567890_abcde"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document processed successfully or validation completed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "status": {
                      "type": "string",
                      "enum": [
                        "document_processed",
                        "document_processed_with_error",
                        "completed_no_changes",
                        "completed_with_changes",
                        "completed_with_pr_created",
                        "completed_with_pr_error"
                      ]
                    },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "currentDocument": { "$ref": "#/components/schemas/DocumentWithPolicies" },
                        "analysis": { "$ref": "#/components/schemas/AnalysisResult" },
                        "progress": {
                          "type": "object",
                          "properties": {
                            "current": { "type": "number" },
                            "total": { "type": "number" },
                            "percentage": { "type": "number" }
                          }
                        },
                        "nextStep": { "type": "string" },
                        "updatedPolicies": { "type": "object" },
                        "pullRequest": {
                          "type": "object",
                          "properties": {
                            "url": { "type": "string" },
                            "number": { "type": "number" }
                          }
                        },
                        "totalChanges": { "type": "number" },
                        "totalPlatformsUpdated": { "type": "number" },
                        "platformsUpdated": {
                          "type": "array",
                          "items": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing validationId parameter."
          },
          "404": {
            "description": "Validation session not found."
          },
          "500": {
            "description": "Internal server error during document processing."
          }
        }
      },
      "options": {
        "summary": "CORS preflight for policy validation processing",
        "responses": {
          "200": {
            "description": "CORS preflight successful."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PolicyUpdate": {
        "type": "object",
        "properties": {
          "reference": { "type": "string" },
          "policy": { "type": "string" },
          "removalCriteria": {
            "type": "array",
            "items": { "type": "string" }
          },
          "evidenceRequirements": {
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["reference", "policy", "removalCriteria", "evidenceRequirements"]
      },
      "DocumentWithPolicies": {
        "type": "object",
        "properties": {
          "platformId": { "type": "string" },
          "platformName": { "type": "string" },
          "reference": { "type": "string" },
          "title": { "type": "string" },
          "url": { "type": "string" },
          "accessTimestamp": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true },
          "policies": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PolicyUpdate" }
          }
        },
        "required": ["platformId", "platformName", "reference", "title", "url", "policies"]
      },
      "AnalysisResult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["no_change", "updated", "error"]
          },
          "reasoning": { "type": "string" },
          "updatedPolicies": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PolicyUpdate" },
            "nullable": true
          },
          "validation": {
            "type": "object",
            "properties": {
              "validationStatus": {
                "type": "string",
                "enum": ["valid", "invalid"]
              },
              "reasoning": { "type": "string" },
              "issues": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "severity": {
                      "type": "string",
                      "enum": ["critical", "minor"]
                    },
                    "type": {
                      "type": "string",
                      "enum": ["hallucination", "meaningless_rewording", "structural_error"]
                    },
                    "description": { "type": "string" }
                  }
                }
              }
            },
            "nullable": true
          }
        },
        "required": ["status", "reasoning"]
      }
    }
  }
}